Solucion de problemas cuando se copiaban los argumentos para un programa con malloc

+++ Kernel/include/pipe_struct.h
@@ -13,23 +13,23 @@ typedef struct{
     char buff[PIPESIZE];
     uint64_t pid_write_lock[MAXLOCK];
     uint64_t pid_read_lock[MAXLOCK];
-    uint8_t count_access;
-    uint8_t index_write;
-    uint8_t index_read;
+    uint64_t count_access;
+    uint64_t index_write;
+    uint64_t index_read;
 }pipe_info;
 
 typedef struct{
     pipe_info * info;
     Pipe_modes mode;
-    uint8_t count_access;
+    uint64_t count_access;
 }pipe_restrict;
 
 typedef struct{
     char * name;
     uint64_t pid_write_lock[MAXLOCK];
     uint64_t pid_read_lock[MAXLOCK];
-    uint8_t index_write;
-    uint8_t index_read;
+    uint64_t index_write;
+    uint64_t index_read;
 }pipe_user_info;
 
 #endif //TP2_SO_PIPE_STRUCT_H
